Added default button layouts and inactive profile constants, as well as methods to replace them in CProfiles, moved helpper functions to the h file, filled out showProfilesFromMemory, misc CProfile improvements, and some printf clarity.
CProfile struct now has a bool for weather the profile is active, uk1 has been added to the id char,  and end now defaults to 0. Functions have also been added to reset the button layout of the profule, or "remove" it, marking it as inactive and removing the name.

Helper functions have been moved to the utc.h file for organizational purposes.

showProfilesFromMemory has been filled out and tested

loadProfileFromConsole still needs to be tested but it's 430am and I have a morning class tomorrow.

TODO reminding me to document functions has been added and will be done at some point.

Default button layouts for all 3 controllers, as well as a deactivated profile have been added as constants.

Next update is filling in more functions, testing, and a demo menu.